+++ Changes.md
@@ -3,6 +3,7 @@ Change Log in version 3.11.0.6(2021081005)
 1. Fix 'Title under logo', please resave the 'Layout responsive' settings 'responsivelogo' and 'responsivecoursetitle'
    - Ref: https://moodle.org/mod/forum/discuss.php?d=431105.
 2. Fix 'One Topic font color change not getting applied' - #264.
+3. Add 'OneTopic active tab colour' - #269.
 
 Change Log in version 3.11.0.5(2021081004)
 ===========================================